修复类型转换bug

+++ server-module/src/main/java/com/server/impl/ContactsAccountServerImpl.java
@@ -68,11 +68,11 @@ public class ContactsAccountServerImpl implements ContactsAccountServer {
             businessOrder.setRemark("新增加普通账户数据，增加期初账户余额" + balance + "元");
         }else{
             //往来账户
-            if(0 > Integer.valueOf(balance)){//如果余额为负数
+            if(0 > Float.valueOf(balance)){//如果余额为负数
                 businessOrder.setDocumentType("1");
                 businessOrder.setExpenditure(contactsAccount.getId());
-                businessOrder.setAmount(String.valueOf(Integer.valueOf(balance) * -1));
-                businessOrder.setRemark("新增加往来账户数据，增加往来金额" + Integer.valueOf(balance) * -1 + "元");
+                businessOrder.setAmount(String.valueOf(Float.valueOf(balance) * -1));
+                businessOrder.setRemark("新增加往来账户数据，增加往来金额" + Float.valueOf(balance) * -1 + "元");
             }else{
                 businessOrder.setDocumentType("2");
                 businessOrder.setIncome(contactsAccount.getId());
